+++ src/Hal/Serializer/ItemNormalizer.php
@@ -67,6 +67,7 @@ final class ItemNormalizer extends AbstractItemNormalizer
 
         $context = $this->initContext($resourceClass, $context);
         $iri = $this->iriConverter->getIriFromResource($object, UrlGeneratorInterface::ABS_PATH, $context['operation'] ?? null, $context);
+
         $context['iri'] = $iri;
         $context['api_normalize'] = true;
 
@@ -75,6 +76,7 @@ final class ItemNormalizer extends AbstractItemNormalizer
         }
 
         $data = parent::normalize($object, $format, $context);
+
         if (!\is_array($data)) {
             return $data;
         }
@@ -143,31 +145,58 @@ final class ItemNormalizer extends AbstractItemNormalizer
         foreach ($attributes as $attribute) {
             $propertyMetadata = $this->propertyMetadataFactory->create($context['resource_class'], $attribute, $options);
 
-            // TODO: 3.0 support multiple types, default value of types will be [] instead of null
-            $type = $propertyMetadata->getBuiltinTypes()[0] ?? null;
-            $isOne = $isMany = false;
+            $types = $propertyMetadata->getBuiltinTypes() ?? [];
 
-            if (null !== $type) {
-                if ($type->isCollection()) {
-                    $valueType = $type->getCollectionValueTypes()[0] ?? null;
-                    $isMany = null !== $valueType && ($className = $valueType->getClassName()) && $this->resourceClassResolver->isResourceClass($className);
-                } else {
-                    $className = $type->getClassName();
-                    $isOne = $className && $this->resourceClassResolver->isResourceClass($className);
+            // prevent declaring $attribute as attribute if it's already declared as relationship
+            $isRelationship = false;
+
+            foreach ($types as $type) {
+                $isOne = $isMany = false;
+
+                if (null !== $type) {
+                    if ($type->isCollection()) {
+                        $valueType = $type->getCollectionValueTypes()[0] ?? null;
+                        $isMany = null !== $valueType && ($className = $valueType->getClassName()) && $this->resourceClassResolver->isResourceClass($className);
+                    } else {
+                        $className = $type->getClassName();
+                        $isOne = $className && $this->resourceClassResolver->isResourceClass($className);
+                    }
                 }
+
+                if (!$isOne && !$isMany) {
+                    // don't declare it as an attribute too quick: maybe the next type is a valid resource
+                    continue;
+                }
+
+                $relation = ['name' => $attribute, 'cardinality' => $isOne ? 'one' : 'many', 'iri' => null, 'operation' => null];
+
+                // if we specify the uriTemplate, generates its value for link definition
+                // @see ApiPlatform\Serializer\AbstractItemNormalizer:getAttributeValue logic for intentional duplicate content
+                if (($className ?? false) && $uriTemplate = $propertyMetadata->getUriTemplate()) {
+                    $childContext = $this->createChildContext($context, $attribute, $format);
+                    unset($childContext['iri'], $childContext['uri_variables'], $childContext['resource_class'], $childContext['operation'], $childContext['operation_name']);
+
+                    $operation = $this->resourceMetadataCollectionFactory->create($className)->getOperation(
+                        operationName: $uriTemplate,
+                        httpOperation: true
+                    );
+
+                    $relation['iri'] = $this->iriConverter->getIriFromResource($object, UrlGeneratorInterface::ABS_PATH, $operation, $childContext);
+                    $relation['operation'] = $operation;
+                }
+
+                if ($propertyMetadata->isReadableLink()) {
+                    $components['embedded'][] = $relation;
+                }
+
+                $components['links'][] = $relation;
+                $isRelationship = true;
             }
 
-            if (!$isOne && !$isMany) {
+            // if all types are not relationships, declare it as an attribute
+            if (!$isRelationship) {
                 $components['states'][] = $attribute;
-                continue;
             }
-
-            $relation = ['name' => $attribute, 'cardinality' => $isOne ? 'one' : 'many'];
-            if ($propertyMetadata->isReadableLink()) {
-                $components['embedded'][] = $relation;
-            }
-
-            $components['links'][] = $relation;
         }
 
         if (false !== $context['cache_key']) {
@@ -194,14 +223,29 @@ final class ItemNormalizer extends AbstractItemNormalizer
                 continue;
             }
 
-            $attributeValue = $this->getAttributeValue($object, $relation['name'], $format, $context);
-            if (empty($attributeValue)) {
-                continue;
-            }
-
             $relationName = $relation['name'];
             if ($this->nameConverter) {
                 $relationName = $this->nameConverter->normalize($relationName, $class, $format, $context);
+            }
+
+            // if we specify the uriTemplate, then the link takes the uriTemplate defined.
+            if ('links' === $type && $iri = $relation['iri']) {
+                $data[$key][$relationName]['href'] = $iri;
+                continue;
+            }
+
+            $childContext = $this->createChildContext($context, $relationName, $format);
+            unset($childContext['iri'], $childContext['uri_variables'], $childContext['operation'], $childContext['operation_name']);
+
+            if ($operation = $relation['operation']) {
+                $childContext['operation'] = $operation;
+                $childContext['operation_name'] = $operation->getName();
+            }
+
+            $attributeValue = $this->getAttributeValue($object, $relation['name'], $format, $childContext);
+
+            if (empty($attributeValue)) {
+                continue;
             }
 
             if ('one' === $relation['cardinality']) {