+++ src/Hydra/Serializer/CollectionFiltersNormalizer.php
@@ -13,11 +13,12 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Hydra\Serializer;
 
-use ApiPlatform\Api\FilterInterface;
 use ApiPlatform\Api\FilterLocatorTrait;
-use ApiPlatform\Api\ResourceClassResolverInterface;
+use ApiPlatform\Doctrine\Odm\State\Options as ODMOptions;
 use ApiPlatform\Doctrine\Orm\State\Options;
+use ApiPlatform\Metadata\FilterInterface;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
+use ApiPlatform\Metadata\ResourceClassResolverInterface;
 use ApiPlatform\Serializer\CacheableSupportsMethodInterface;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Serializer\Exception\UnexpectedValueException;
@@ -99,6 +100,7 @@ final class CollectionFiltersNormalizer implements NormalizerInterface, Normaliz
         if (!$resourceFilters) {
             return $data;
         }
+
         $requestParts = parse_url($context['request_uri'] ?? '');
         if (!\is_array($requestParts)) {
             return $data;
@@ -110,8 +112,14 @@ final class CollectionFiltersNormalizer implements NormalizerInterface, Normaliz
             }
         }
 
-        if (($options = $operation?->getStateOptions()) && $options instanceof Options && $options->getEntityClass()) {
-            $resourceClass = $options->getEntityClass();
+        if ($options = $operation->getStateOptions()) {
+            if ($options instanceof Options && $options->getEntityClass()) {
+                $resourceClass = $options->getEntityClass();
+            }
+
+            if ($options instanceof ODMOptions && $options->getDocumentClass()) {
+                $resourceClass = $options->getDocumentClass();
+            }
         }
 
         if ($currentFilters) {