+++ src/Symfony/EventListener/DeserializeListener.php
@@ -14,12 +14,11 @@ declare(strict_types=1);
 namespace ApiPlatform\Symfony\EventListener;
 
 use ApiPlatform\Api\FormatMatcher;
-use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
 use ApiPlatform\Serializer\SerializerContextBuilderInterface;
+use ApiPlatform\State\Util\OperationRequestInitiatorTrait;
+use ApiPlatform\Symfony\Util\RequestAttributesExtractor;
 use ApiPlatform\Symfony\Validator\Exception\ValidationException;
-use ApiPlatform\Util\OperationRequestInitiatorTrait;
-use ApiPlatform\Util\RequestAttributesExtractor;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Exception\UnsupportedMediaTypeHttpException;
@@ -77,6 +76,10 @@ final class DeserializeListener
 
         $operation = $this->initializeOperation($request);
 
+        if ('api_platform.symfony.main_controller' === $operation?->getController()) {
+            return;
+        }
+
         if (!($operation?->canDeserialize() ?? true)) {
             return;
         }
@@ -88,9 +91,9 @@ final class DeserializeListener
         if (
             null !== $data
             && (
-                HttpOperation::METHOD_POST === $method
-                || HttpOperation::METHOD_PATCH === $method
-                || (HttpOperation::METHOD_PUT === $method && !($operation->getExtraProperties()['standard_put'] ?? false))
+                'POST' === $method
+                || 'PATCH' === $method
+                || ('PUT' === $method && !($operation->getExtraProperties()['standard_put'] ?? false))
             )
         ) {
             $context[AbstractNormalizer::OBJECT_TO_POPULATE] = $data;