+++ src/Symfony/EventListener/QueryParameterValidateListener.php
@@ -14,12 +14,13 @@ declare(strict_types=1);
 namespace ApiPlatform\Symfony\EventListener;
 
 use ApiPlatform\Api\QueryParameterValidator\QueryParameterValidator;
+use ApiPlatform\Doctrine\Odm\State\Options as ODMOptions;
 use ApiPlatform\Doctrine\Orm\State\Options;
 use ApiPlatform\Metadata\CollectionOperationInterface;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
-use ApiPlatform\Util\OperationRequestInitiatorTrait;
-use ApiPlatform\Util\RequestAttributesExtractor;
-use ApiPlatform\Util\RequestParser;
+use ApiPlatform\State\Util\OperationRequestInitiatorTrait;
+use ApiPlatform\State\Util\RequestParser;
+use ApiPlatform\Symfony\Util\RequestAttributesExtractor;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 
 /**
@@ -51,6 +52,9 @@ final class QueryParameterValidateListener
         }
 
         $operation = $this->initializeOperation($request);
+        if ('api_platform.symfony.main_controller' === $operation?->getController()) {
+            return;
+        }
 
         if (!($operation?->getQueryParameterValidationEnabled() ?? true) || !$operation instanceof CollectionOperationInterface) {
             return;
@@ -61,8 +65,14 @@ final class QueryParameterValidateListener
 
         $class = $attributes['resource_class'];
 
-        if (($options = $operation->getStateOptions()) && $options instanceof Options && $options->getEntityClass()) {
-            $class = $options->getEntityClass();
+        if ($options = $operation->getStateOptions()) {
+            if ($options instanceof Options && $options->getEntityClass()) {
+                $class = $options->getEntityClass();
+            }
+
+            if ($options instanceof ODMOptions && $options->getDocumentClass()) {
+                $class = $options->getDocumentClass();
+            }
         }
 
         $this->queryParameterValidator->validateFilters($class, $operation->getFilters() ?? [], $queryParameters);