+++ src/Symfony/EventListener/ReadListener.php
@@ -16,17 +16,16 @@ namespace ApiPlatform\Symfony\EventListener;
 use ApiPlatform\Api\UriVariablesConverterInterface;
 use ApiPlatform\Exception\InvalidIdentifierException;
 use ApiPlatform\Exception\InvalidUriVariableException;
-use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Put;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
+use ApiPlatform\Metadata\Util\CloneTrait;
 use ApiPlatform\Serializer\SerializerContextBuilderInterface;
 use ApiPlatform\State\Exception\ProviderNotFoundException;
 use ApiPlatform\State\ProviderInterface;
 use ApiPlatform\State\UriVariablesResolverTrait;
-use ApiPlatform\Util\CloneTrait;
-use ApiPlatform\Util\OperationRequestInitiatorTrait;
-use ApiPlatform\Util\RequestAttributesExtractor;
-use ApiPlatform\Util\RequestParser;
+use ApiPlatform\State\Util\OperationRequestInitiatorTrait;
+use ApiPlatform\State\Util\RequestParser;
+use ApiPlatform\Symfony\Util\RequestAttributesExtractor;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
@@ -61,6 +60,10 @@ final class ReadListener
         $request = $event->getRequest();
         $operation = $this->initializeOperation($request);
 
+        if ('api_platform.symfony.main_controller' === $operation?->getController()) {
+            return;
+        }
+
         if (!($attributes = RequestAttributesExtractor::extractAttributes($request))) {
             return;
         }
@@ -90,7 +93,12 @@ final class ReadListener
         $resourceClass = $operation->getClass() ?? $attributes['resource_class'];
         try {
             $uriVariables = $this->getOperationUriVariables($operation, $parameters, $resourceClass);
-            $data = $this->provider->provide($operation, $uriVariables, $context);
+            if ($request->attributes->get('_api_error', false)) {
+                $exception = $request->attributes->get('data');
+                $data = $operation->getProvider() ? $this->provider->provide($operation, $uriVariables, $context + ['previous_data' => $exception]) : $exception;
+            } else {
+                $data = $this->provider->provide($operation, $uriVariables, $context);
+            }
         } catch (InvalidIdentifierException|InvalidUriVariableException $e) {
             throw new NotFoundHttpException('Invalid identifier value or configuration.', $e);
         } catch (ProviderNotFoundException $e) {
@@ -99,9 +107,9 @@ final class ReadListener
 
         if (
             null === $data
-            && HttpOperation::METHOD_POST !== $operation->getMethod()
+            && 'POST' !== $operation->getMethod()
             && (
-                HttpOperation::METHOD_PUT !== $operation->getMethod()
+                'PUT' !== $operation->getMethod()
                 || ($operation instanceof Put && !($operation->getAllowCreate() ?? false))
             )
         ) {