+++ tests/Symfony/Bundle/DependencyInjection/ConfigurationTest.php
@@ -82,9 +82,12 @@ class ConfigurationTest extends TestCase
             'description' => 'description',
             'version' => '1.0.0',
             'show_webby' => true,
-            'formats' => [
-                'jsonld' => ['mime_types' => ['application/ld+json']],
+            'formats' => [],
+            'docs_formats' => [
+                'jsonopenapi' => ['mime_types' => ['application/vnd.openapi+json']],
+                'yamlopenapi' => ['mime_types' => ['application/vnd.openapi+yaml']],
                 'json' => ['mime_types' => ['application/json']],
+                'jsonld' => ['mime_types' => ['application/ld+json']],
                 'html' => ['mime_types' => ['text/html']],
             ],
             'patch_formats' => [
@@ -119,6 +122,9 @@ class ConfigurationTest extends TestCase
                     'enabled' => true,
                 ],
                 'graphiql' => [
+                    'enabled' => true,
+                ],
+                'introspection' => [
                     'enabled' => true,
                 ],
                 'nesting_separator' => '_',
@@ -217,6 +223,8 @@ class ConfigurationTest extends TestCase
             'maker' => [
                 'enabled' => true,
             ],
+            'keep_legacy_inflector' => true,
+            'event_listeners_backward_compatibility_layer' => true,
         ], $config);
     }
 
@@ -279,6 +287,26 @@ class ConfigurationTest extends TestCase
     /**
      * Test config for api keys.
      */
+    public function testInvalidApiKeysConfig(): void
+    {
+        $this->expectExceptionMessage('The api keys "key" is not valid according to the pattern enforced by OpenAPI 3.1 ^[a-zA-Z0-9._-]+$.');
+        $exampleConfig = [
+            'name' => 'Authorization',
+            'type' => 'query',
+        ];
+
+        $config = $this->processor->processConfiguration($this->configuration, [
+            'api_platform' => [
+                'swagger' => [
+                    'api_keys' => ['Some Authorization name, like JWT' => $exampleConfig, 'Another-Auth' => $exampleConfig],
+                ],
+            ],
+        ]);
+    }
+
+    /**
+     * Test config for api keys.
+     */
     public function testApiKeysConfig(): void
     {
         $exampleConfig = [
@@ -289,13 +317,13 @@ class ConfigurationTest extends TestCase
         $config = $this->processor->processConfiguration($this->configuration, [
             'api_platform' => [
                 'swagger' => [
-                    'api_keys' => ['Some Authorization name, like JWT' => $exampleConfig],
+                    'api_keys' => ['authorization_name_like_JWT' => $exampleConfig],
                 ],
             ],
         ]);
 
         $this->assertArrayHasKey('api_keys', $config['swagger']);
-        $this->assertSame($exampleConfig, $config['swagger']['api_keys']['Some Authorization name, like JWT']);
+        $this->assertSame($exampleConfig, $config['swagger']['api_keys']['authorization_name_like_JWT']);
     }
 
     /**